Move to webpack/encore to organize assets
