--- public/css/importFromCsvApp.css
@@ -1,170 +1,0 @@
-/** * Operations */
-a.renderAppAction > img {
-    width: 16px;
-    aspect-ratio: auto 1;
-    height: 16px;
-}
-[v-cloak], .ifcb-import-from-csv-app.ifcb-status-preparing {
-    display: none;
-}
-.ifcb-status-ready .ifcb-progress-bar-container, .ifcb-status-ready .ifcb-import-report-box {
-    visibility: hidden;
-}
-/** * Fields */
-.ifcb-sql-descr {
-    color: #999;
-}
-/** * Detailview */
-.ifcb-file-content-box {
-    max-height: 300px;
-    overflow-x: hidden;
-    padding: 5px;
-    font-family: SFMono-Regular, Menlo, Monaco, Consolas, "Liberation Mono", "Courier New", monospace;
-    background-color: #000;
-    color: #fb77ec;
-}
-.ifcb-file-content-box ul {
-    list-style-type: none;
-}
-.ifcb-file-content-box li code {
-    font-size: 11px;
-    line-height: 15px;
-    white-space: pre-wrap;
-    word-break: break-word;
-}
-.ifcb-m-1 {
-    margin: 15px;
-}
-.ifcb-mt-1 {
-    margin-top: 15px;
-}
-.ifcb-me-1 {
-    margin-right: 15px;
-}
-.ifcb-mb-1 {
-    margin-bottom: 15px;
-}
-.ifcb-ms-1 {
-    margin-left: 15px;
-}
-.ifcb-d-flex {
-    display: flex;
-}
-.ifcb-d-block {
-    display: block;
-}
-.ifcb-brand-logo {
-    width: 150px;
-}
-.ifcb-help-text a {
-    color: blue;
-}
-.ifcb-custom-header-back, .ifcb-edit-header {
-    padding: 3px 0 3px 21px;
-    background-position: left center;
-    background-repeat: no-repeat;
-    margin-left: 15px;
-}
-.ifcb-custom-header-back {
-    background-image: url('../../../system/themes/flexible/icons/back.svg');
-}
-.ifcb-edit-header {
-    background-image: url('../../../system/themes/flexible/icons/edit.svg');
-}
-.ifcb-import-from-csv-app .widget {
-    margin: 15px;
-}
-.ifcb-import-from-csv-app .ifcb-progress-bar-container {
-    margin: 15px 0;
-    background-color: #fbfbfb;
-}
-.ifcb-import-from-csv-app .ifcb-progress-bar-container .ifcb-progress-bar-inner {
-    height: 30px;
-    width: 0;
-    background-color: #5eaf4c;
-    transition: width 1.5s ease;
-    line-height: 1.5;
-    color: #fff;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
-.ifcb-import-from-csv-app button.tl_submit.ifcb-display-errors-btn {
-    background-color: red;
-    border: 1px solid red !important;
-    color: #fff;
-    padding: 8px;
-    height: auto;
-}
-.ifcb-import-from-csv-app button.tl_submit.ifcb-display-errors-btn:hover {
-    background-color: #900;
-    border-color: #900 !important;
-}
-.ifcb-import-from-csv-app button.tl_submit.ifcb-display-errors-btn.hideSuccess {
-    background-color: #5eaf4c;
-    color: #fff;
-    border-color: #5eaf4c !important;
-}
-.ifcb-import-from-csv-app .ifcb-import-report-box {
-    margin: 15px 0;
-    padding: 15px;
-    max-height: 400px;
-    overflow-y: scroll;
-    background-color: #000;
-}
-.ifcb-import-from-csv-app .ifcb-import-report-box .ifcb-import-report-table {
-    width: 100%;
-}
-.ifcb-import-from-csv-app .ifcb-import-report-box .ifcb-import-report-table * {
-    font-weight: normal !important;
-    font-size: 10px !important;
-}
-.ifcb-import-from-csv-app .ifcb-import-report-box .ifcb-import-report-table td {
-    border: 0;
-    padding: 4px;
-    font-size: 10px;
-    width: 50%;
-}
-.ifcb-import-from-csv-app .ifcb-import-report-box .ifcb-import-report-table td.ifcb-td-title {
-    font-weight: bold;
-}
-.ifcb-import-from-csv-app .ifcb-import-report-box .ifcb-import-report-table tr.ifcb-import-success td {
-    color: #7dc06e !important;
-}
-.ifcb-import-from-csv-app .ifcb-import-report-box .ifcb-import-report-table tr.ifcb-import-failure td {
-    color: #ff7d7d !important;
-}
-.ifcb-import-from-csv-app .ifcb-import-report-box .ifcb-import-report-table tr.ifcb-delim td {
-    border-left-width: 0;
-    border-right-width: 0;
-}
-.ifcb-import-from-csv-app .ifcb-import-report-box .ifcb-import-report-table tr.hiddenRow, .ifcb-import-from-csv-app .ifcb-import-report-box .ifcb-import-report-table tr.hiddenRow td {
-    display: none;
-}
-.ifcb-import-from-csv-app .ifcb-import-report-box .ifcb-import-report-table tr.ifcb-delim td {
-    border: none;
-}
-.ifcb-import-from-csv-app .ifcb-status-box {
-    min-height: 30px;
-    color: #b4b4b4;
-}
-.ifcb-import-from-csv-app .ifcb-import-summary {
-    margin: 16px 0;
-    line-height: 1.5;
-}
-.ifcb-import-from-csv-app .ifcb-import-summary li.ifcb-import-failure {
-    color: red !important;
-}
-.ifcb-import-from-csv-app .ifcb-import-summary li.ifcb-import-success {
-    color: #008000 !important;
-}
-.ifcb-import-from-csv-app .tl_submit_container button:not([disabled]).ifcb-import-button {
-    color: #fff;
-    background: red;
-    border-color: red;
-}
-.ifcb-import-from-csv-app .tl_submit_container button:not([disabled]).ifcb-import-button:hover {
-    background: #b30000;
-    border-color: #b30000;
-}
-
