+++ public/css/import_from_csv_app.06a92088.css
@@ -1,0 +1,2 @@
+a.renderAppAction>img{aspect-ratio:auto 16/16;height:16px;width:16px}.ifcb-import-from-csv-app.ifcb-status-preparing,[v-cloak]{display:none}.ifcb-status-ready .ifcb-import-report-box,.ifcb-status-ready .ifcb-progress-bar-container{visibility:hidden}.ifcb-sql-descr{color:#999}.ifcb-file-content-box{background-color:#000;color:#fb77ec;font-family:SFMono-Regular,Menlo,Monaco,Consolas,Liberation Mono,Courier New,monospace;max-height:300px;overflow-x:hidden;padding:5px}.ifcb-file-content-box ul{list-style-type:none}.ifcb-file-content-box li code{font-size:11px;line-height:15px;white-space:pre-wrap;word-break:break-word}.ifcb-m-1{margin:15px}.ifcb-mt-1{margin-top:15px}.ifcb-me-1{margin-right:15px}.ifcb-mb-1{margin-bottom:15px}.ifcb-ms-1{margin-left:15px}.ifcb-d-flex{display:flex}.ifcb-d-block{display:block}.ifcb-brand-logo{width:150px}.ifcb-help-text a{color:blue}.ifcb-custom-header-back,.ifcb-edit-header{background-position:0;background-repeat:no-repeat;margin-left:15px;padding:3px 0 3px 21px}.ifcb-custom-header-back{background-image:url(/bundles/markocupicimportfromcsv/images/back.054a6833.svg)}.ifcb-edit-header{background-image:url(/bundles/markocupicimportfromcsv/images/edit.0b1829c8.svg)}.ifcb-import-from-csv-app .widget{margin:15px}.ifcb-import-from-csv-app .ifcb-progress-bar-container{background-color:#fbfbfb;margin:15px 0}.ifcb-import-from-csv-app .ifcb-progress-bar-container .ifcb-progress-bar-inner{align-items:center;background-color:#5eaf4c;color:#fff;display:flex;height:30px;justify-content:center;line-height:1.5;transition:width 1.5s ease;width:0}.ifcb-import-from-csv-app button.tl_submit.ifcb-display-errors-btn{background-color:red;border:1px solid red!important;color:#fff;height:auto;padding:8px}.ifcb-import-from-csv-app button.tl_submit.ifcb-display-errors-btn:hover{background-color:#c00;border-color:#c00}.ifcb-import-from-csv-app button.tl_submit.ifcb-display-errors-btn.hideSuccess{background-color:#5eaf4c;border-color:#5eaf4c!important;color:#fff}.ifcb-import-from-csv-app .ifcb-import-report-box{background-color:#000;margin:15px 0;max-height:400px;overflow-y:scroll;padding:15px}.ifcb-import-from-csv-app .ifcb-import-report-box .ifcb-import-report-table{width:100%}.ifcb-import-from-csv-app .ifcb-import-report-box .ifcb-import-report-table *{font-size:10px!important;font-weight:400!important}.ifcb-import-from-csv-app .ifcb-import-report-box .ifcb-import-report-table td{border:0;font-size:10px;padding:4px;width:50%}.ifcb-import-from-csv-app .ifcb-import-report-box .ifcb-import-report-table td.ifcb-td-title{font-weight:700}.ifcb-import-from-csv-app .ifcb-import-report-box .ifcb-import-report-table tr.ifcb-import-success td{color:#6db85c!important}.ifcb-import-from-csv-app .ifcb-import-report-box .ifcb-import-report-table tr.ifcb-import-failure td{color:#ff7d7d!important}.ifcb-import-from-csv-app .ifcb-import-report-box .ifcb-import-report-table tr.ifcb-delim td{border-left-width:0;border-right-width:0}.ifcb-import-from-csv-app .ifcb-import-report-box .ifcb-import-report-table tr.hiddenRow,.ifcb-import-from-csv-app .ifcb-import-report-box .ifcb-import-report-table tr.hiddenRow td{display:none}.ifcb-import-from-csv-app .ifcb-import-report-box .ifcb-import-report-table tr.ifcb-delim td{border:none}.ifcb-import-from-csv-app .ifcb-status-box{color:#b4b4b4;min-height:30px}.ifcb-import-from-csv-app .ifcb-import-summary{line-height:1.5;margin:16px 0}.ifcb-import-from-csv-app .ifcb-import-summary li.ifcb-import-failure{color:red!important}.ifcb-import-from-csv-app .ifcb-import-summary li.ifcb-import-success{color:green!important}.ifcb-import-from-csv-app .tl_submit_container button:not([disabled]).ifcb-import-button{background:red;border-color:red;color:#fff}.ifcb-import-from-csv-app .tl_submit_container button:not([disabled]).ifcb-import-button:hover{background:#d90000;border-color:#d90000}
+/*# sourceMappingURL=import_from_csv_app.06a92088.css.map*/
