--- public/css/loader.css
@@ -1,59 +1,0 @@
-.ifcb-flex-center {
-  display: flex;
-  justify-content: center;
-}
-
-.ifcb-flex-column {
-  display: flex;
-  flex-direction: column;
-}
-
-.ifcb-loader-box {
-  height: 210px;
-  color: #999;
-  padding-top: 100px;
-}
-
-.ifcb-loader {
-  display: inline-block;
-  position: relative;
-  width: 80px;
-  height: 80px;
-}
-
-.ifcb-loader div {
-  display: inline-block;
-  position: absolute;
-  left: 8px;
-  width: 16px;
-  background: #eeeeee;
-  animation: ifcb-loader 1.2s cubic-bezier(0, 0.5, 0.5, 1) infinite;
-}
-
-.ifcb-loader div:nth-child(1) {
-  left: 8px;
-  animation-delay: -0.24s;
-}
-
-.ifcb-loader div:nth-child(2) {
-  left: 32px;
-  animation-delay: -0.12s;
-}
-
-.ifcb-loader div:nth-child(3) {
-  left: 56px;
-  animation-delay: 0;
-}
-
-@keyframes ifcb-loader {
-  0% {
-    top: 8px;
-    height: 64px;
-  }
-  50%, 100% {
-    top: 24px;
-    height: 32px;
-  }
-}
-
-/*# sourceMappingURL=loader.css.map */
