+++ public/css/loader.e2aad614.css
@@ -1,0 +1,2 @@
+.ifcb-flex-center{display:flex;justify-content:center}.ifcb-flex-column{display:flex;flex-direction:column}.ifcb-loader-box{color:#999;height:210px;padding-top:100px}.ifcb-loader{display:inline-block;height:80px;position:relative;width:80px}.ifcb-loader div{animation:ifcb-loader 1.2s cubic-bezier(0,.5,.5,1) infinite;background:#eee;display:inline-block;left:8px;position:absolute;width:16px}.ifcb-loader div:first-child{animation-delay:-.24s;left:8px}.ifcb-loader div:nth-child(2){animation-delay:-.12s;left:32px}.ifcb-loader div:nth-child(3){animation-delay:0;left:56px}@keyframes ifcb-loader{0%{height:64px;top:8px}50%,to{height:32px;top:24px}}
+/*# sourceMappingURL=loader.e2aad614.css.map*/
