+++ public/js/import_from_csv_app.7f21985b.js
@@ -1,0 +1,2 @@
+/*! For license information please see import_from_csv_app.7f21985b.js.LICENSE.txt */
+class ImportFromCsvApp{constructor(t,e){const{createApp:s}=Vue,r=s({data(){return{items:[],isTestMode:!1,status:"ifcb-status-preparing",urlStack:[],requestsDone:0,requestsNeeded:0,model:{urls:[],count:0,limit:0,offset:0},head:{countSuccess:0,countErrors:0,countTotal:0},report:{logs:[],summary:{rows:0,errors:0,success:0}}}},watch:{report:{handler(t,e){const s=document.getElementById("ifcbImportReportBox");s&&window.setTimeout(()=>s.scrollTop=s.scrollHeight+100,100)},deep:!0}},mounted(){this.$nextTick(function(){this.appMountAction(e)})},methods:{appMountAction(t){fetch(t,{method:"GET",headers:{"x-requested-with":"XMLHttpRequest"}}).then(t=>(200===t.status&&t.json().then(t=>{Object.keys(t.data.model).forEach(e=>{this.model[e]=t.data.model[e]}),this.urlStack=t.data.urlStack,this.requestsNeeded=this.urlStack.length}),t)).then(t=>{setTimeout(()=>this.status="ifcb-status-ready",2e3)}).catch(t=>{console.error("There was en error: "+t)})},importAction(t,e=!1){if(!0===e&&"ifcb-status-ready"!==this.status)return void console.error("The import cannot be called twice.");this.disableButtons(),!0===t&&(this.isTestMode=!0);let s="";this.urlStack.length?(this.status="ifcb-status-pending",s=this.urlStack.shift(),s=s.replace("_isTestMode_",!0===t?"true":"false"),fetch(s,{method:"GET",headers:{"x-requested-with":"XMLHttpRequest"}}).then(t=>{if(200===t.status)return t.json().then(t=>{this.report.logs=[...this.report.logs,...t.data.logs],this.report.summary.rows=this.report.summary.rows+t.data.summary.rows,this.report.summary.success=this.report.summary.success+t.data.summary.success,this.report.summary.errors=this.report.summary.errors+t.data.summary.errors,this.requestsDone++,this.updateProgressBar()}),t;this.status="ifcb-status-error"}).then(e=>{setTimeout(()=>this.importAction(t,!1),200)}).catch(t=>{console.error("There was en error: "+t)})):this.status="ifcb-status-completed"},disableButtons(){let t=document.querySelectorAll("#ifcbImportFromCsvApp .tl_submit_container button");t&&t.forEach(function(t){t.setAttribute("disabled","true")})},filterInsertListing(t){const e=t.target;e.classList.toggle("hideSuccess"),e.classList.contains("hideSuccess")?e.innerText=e.getAttribute("data-lbl-all"):e.innerText=e.getAttribute("data-lbl-error-only");let s=document.querySelectorAll("tr.ifcb-import-success,tr.ifcb-delim");s&&s.forEach(t=>{t.classList.toggle("hiddenRow")})},updateProgressBar(){const t=document.getElementById("importProgress"),e=document.querySelector("#importProgress .ifcb-percentage");if(t){let s=Math.ceil((this.requestsNeeded-this.urlStack.length)/this.requestsNeeded*100);if(0===this.requestsNeeded){}t.style.width=s+"%",e&&(e.innerHTML=s+" %")}}}});return r.config.compilerOptions.delimiters=["${","}"],r.mount(t)}}
