+++ public/js/import_from_csv_app.7f21985b.js.LICENSE.txt
@@ -1,0 +1,9 @@
+/*
+ * This file is part of Import From CSV Bundle.
+ *
+ * (c) Marko Cupic <m.cupic@gmx.ch>
+ * @license MIT
+ * For the full copyright and license information,
+ * please view the LICENSE file that was distributed with this source code.
+ * @link https://github.com/markocupic/import-from-csv-bundle
+ */
