+++ public/manifest.json
@@ -1,0 +1,12 @@
+{
+  "css/import_from_csv_app.css": "/bundles/markocupicimportfromcsv/css/import_from_csv_app.06a92088.css",
+  "css/loader.css": "/bundles/markocupicimportfromcsv/css/loader.e2aad614.css",
+  "js/vue/dist/vue.global.prod.js": "/bundles/markocupicimportfromcsv/js/vue/dist/vue.global.prod.3036993a.js",
+  "images/logo.png": "/bundles/markocupicimportfromcsv/images/logo.afcbbd7e.png",
+  "js/import_from_csv_app.js": "/bundles/markocupicimportfromcsv/js/import_from_csv_app.7f21985b.js",
+  "images/import.svg": "/bundles/markocupicimportfromcsv/images/import.70ec8b2e.svg",
+  "images/edit.svg": "/bundles/markocupicimportfromcsv/images/edit.0b1829c8.svg",
+  "images/back.svg": "/bundles/markocupicimportfromcsv/images/back.054a6833.svg",
+  "import_from_csv_app.06a92088.css.map": "/bundles/markocupicimportfromcsv/css/import_from_csv_app.06a92088.css.map",
+  "loader.e2aad614.css.map": "/bundles/markocupicimportfromcsv/css/loader.e2aad614.css.map"
+}
