+++ templates/help_text.html.twig
@@ -1,10 +1,9 @@
 <div class="widget ifcb-manual ifcb-d-block">
     <div class="ifcb-mb-1">
         <a href="https://github.com/markocupic/import-from-csv-bundle">
-            <img class="ifcb-brand-logo" src="bundles/markocupicimportfromcsv/logo.png">
+            <img class="ifcb-brand-logo" src="{{ asset('images/logo.png', 'markocupic_import_from_csv') }}">
         </a>
     </div>
-
     <div class="ifcb-help-text">
         {{ help_text|raw }}
     </div>
